+++ packages/core/src/document/pdf-date.ts
